Reduce the levels of timeouts in execute timeout test
Issue #2175 found that execute resource timeouts weren't working on Windows
but we were waiting 600 seconds for the test to fail. #2686 added a Timeout
to make sure we didn't run too long, but this option simplifies the test by
reducing the number of timeouts involved.

We're currently seeing intermittant failures on a FreeBSD tester, which is
likely a performance issue. But it's possible that we're seeing another issue
that we don't yet understand, so I don't want to throw away the regression
test yet.